Updating startup to use ready instead of load if no preloader is defined
Updating startup logic to use .ready() instead of .bind('load') if no preloader is defined. 

The issue here is two-fold:
1. It's possible that sequence's initialization will happen after the load event has already fired (such as if loading asynchronously). 
2. The load event can be unreliable and it's usually best practice to use the jquery ready() event instead.

If the user really wants to wait for images to load they can make use of the preloader or initialize sequence manually when needed.

Note: this change may encourage the renaming of the 'oncePreloaded()' function.